feat: mosaic wizard UI refresh with 2-step flow and thumbnail cards
Decompose the 1114-line monolithic MosaicWizard into a clean 2-step
wizard matching the Composite Wizard patterns:

- Extract FootprintPreview.tsx from inline component
- Create MosaicSelectStep with thumbnail card grid, search/filter
  dropdowns, target grouping, and inline footprint preview
- Create MosaicPreviewStep merging old Configure + Generate steps
  with grid layout (preview left, settings sidebar right)
- Rewrite MosaicWizard.tsx as thin ~200-line shell using WizardStepper
- Simplify MosaicWizard.css to shared shell + footprint styles
- Update MosaicTypes.ts step type from 1|2|3 to 1|2
- Debounced footprint loading on selection change (500ms)
