+++ frontend/jwst-frontend/src/components/wizard/MosaicSelectStep.css
@@ -1,0 +1,315 @@
+/* MosaicSelectStep — thumbnail card selection with search/filter/target grouping */
+
+.mosaic-select-step {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  overflow: hidden;
+  min-height: 0;
+}
+
+/* Toolbar: search + filters */
+.mosaic-select-toolbar {
+  flex-shrink: 0;
+  padding: 1.25rem 1.5rem;
+  border-bottom: 1px solid rgba(74, 144, 217, 0.15);
+}
+
+.mosaic-search-bar {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  margin-bottom: 0.75rem;
+}
+
+.mosaic-search-input {
+  flex: 1;
+  padding: 0.6rem 1rem;
+  background: rgba(0, 0, 0, 0.3);
+  border: 1px solid rgba(74, 144, 217, 0.2);
+  border-radius: 8px;
+  color: #e5e7eb;
+  font-size: 0.9rem;
+  outline: none;
+  transition: border-color 0.2s;
+}
+
+.mosaic-search-input:focus {
+  border-color: rgba(74, 144, 217, 0.5);
+}
+
+.mosaic-search-input::placeholder {
+  color: #6b7280;
+}
+
+.mosaic-selection-count {
+  color: #4a90d9;
+  font-size: 0.85rem;
+  font-weight: 500;
+  white-space: nowrap;
+}
+
+.mosaic-selection-actions {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.mosaic-selection-btn {
+  padding: 0.4rem 0.65rem;
+  border-radius: 6px;
+  border: 1px solid rgba(74, 144, 217, 0.3);
+  background: rgba(74, 144, 217, 0.15);
+  color: #9dc9f0;
+  font-size: 0.75rem;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  white-space: nowrap;
+}
+
+.mosaic-selection-btn:hover:not(:disabled) {
+  background: rgba(74, 144, 217, 0.3);
+  color: #d4e9fb;
+}
+
+.mosaic-selection-btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.mosaic-filter-row {
+  display: grid;
+  grid-template-columns: repeat(3, minmax(0, 1fr));
+  gap: 0.75rem;
+}
+
+.mosaic-filter-control {
+  display: flex;
+  flex-direction: column;
+  gap: 0.35rem;
+}
+
+.mosaic-filter-control label {
+  color: #9ca3af;
+  font-size: 0.78rem;
+  font-weight: 500;
+}
+
+.mosaic-filter-control select {
+  width: 100%;
+  padding: 0.55rem 0.6rem;
+  background: rgba(0, 0, 0, 0.3);
+  border: 1px solid rgba(74, 144, 217, 0.2);
+  border-radius: 6px;
+  color: #e5e7eb;
+  font-size: 0.83rem;
+  outline: none;
+}
+
+.mosaic-filter-control select:focus {
+  border-color: rgba(74, 144, 217, 0.5);
+}
+
+.mosaic-filter-control select option {
+  background: #1a1a2e;
+  color: #e5e7eb;
+}
+
+/* Scrollable card grid area */
+.mosaic-card-grid-scroll {
+  flex: 1;
+  overflow-y: auto;
+  padding: 1rem 1.5rem;
+  min-height: 0;
+}
+
+.mosaic-card-grid-scroll::-webkit-scrollbar {
+  width: 6px;
+}
+
+.mosaic-card-grid-scroll::-webkit-scrollbar-track {
+  background: rgba(0, 0, 0, 0.2);
+  border-radius: 3px;
+}
+
+.mosaic-card-grid-scroll::-webkit-scrollbar-thumb {
+  background: rgba(74, 144, 217, 0.3);
+  border-radius: 3px;
+}
+
+.mosaic-empty {
+  color: #6b7280;
+  text-align: center;
+  padding: 2rem;
+}
+
+/* Target group sections */
+.mosaic-target-group {
+  margin-bottom: 1.25rem;
+}
+
+.mosaic-target-group:last-child {
+  margin-bottom: 0;
+}
+
+.mosaic-target-header {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  margin-bottom: 0.625rem;
+  padding-bottom: 0.375rem;
+  border-bottom: 1px solid rgba(74, 144, 217, 0.12);
+}
+
+.mosaic-target-name {
+  font-size: 0.85rem;
+  font-weight: 600;
+  color: #e5e7eb;
+}
+
+.mosaic-target-count {
+  font-size: 0.72rem;
+  color: #6b7280;
+}
+
+/* Card grid */
+.mosaic-card-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(120px, 1fr));
+  gap: 0.625rem;
+}
+
+/* Image card */
+.mosaic-image-card {
+  display: flex;
+  flex-direction: column;
+  background: rgba(26, 26, 46, 0.8);
+  border: 2px solid rgba(74, 144, 217, 0.15);
+  border-radius: 10px;
+  overflow: hidden;
+  cursor: pointer;
+  transition: all 0.15s ease;
+  user-select: none;
+}
+
+.mosaic-image-card:hover {
+  border-color: rgba(74, 144, 217, 0.4);
+  background: rgba(26, 26, 46, 0.95);
+  transform: translateY(-1px);
+}
+
+.mosaic-image-card.selected {
+  border-color: #4a90d9;
+  background: rgba(74, 144, 217, 0.12);
+  box-shadow: 0 0 0 1px rgba(74, 144, 217, 0.3);
+}
+
+.mosaic-image-card:focus-visible {
+  outline: 2px solid #4a90d9;
+  outline-offset: 2px;
+}
+
+/* Card thumbnail */
+.mosaic-card-thumbnail {
+  position: relative;
+  width: 100%;
+  aspect-ratio: 1;
+  background: #0d1117;
+  overflow: hidden;
+}
+
+.mosaic-card-thumbnail img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.mosaic-card-thumbnail-fallback {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #6b7280;
+}
+
+/* Selection checkmark overlay */
+.mosaic-card-check {
+  position: absolute;
+  top: 6px;
+  right: 6px;
+  width: 24px;
+  height: 24px;
+  border-radius: 50%;
+  background: #4a90d9;
+  color: white;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.4);
+}
+
+/* Card info row */
+.mosaic-card-info {
+  display: flex;
+  align-items: center;
+  gap: 0.375rem;
+  padding: 0.5rem 0.625rem;
+  min-width: 0;
+}
+
+.mosaic-card-filter {
+  flex: 1;
+  font-size: 0.72rem;
+  color: #9ca3af;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.mosaic-card-level {
+  flex-shrink: 0;
+  padding: 0.1rem 0.35rem;
+  border-radius: 3px;
+  color: white;
+  font-size: 0.6rem;
+  font-weight: 600;
+  letter-spacing: 0.02em;
+}
+
+/* Inline footprint preview */
+.mosaic-inline-footprint {
+  flex-shrink: 0;
+  padding: 0.75rem 1.5rem 1rem;
+  border-top: 1px solid rgba(74, 144, 217, 0.15);
+  max-height: 280px;
+  overflow-y: auto;
+}
+
+.mosaic-inline-footprint-title {
+  margin: 0 0 0.5rem;
+  font-size: 0.85rem;
+  font-weight: 600;
+  color: #e5e7eb;
+}
+
+/* Responsive */
+@media (max-width: 768px) {
+  .mosaic-search-bar {
+    flex-wrap: wrap;
+    align-items: stretch;
+  }
+
+  .mosaic-selection-actions {
+    width: 100%;
+    justify-content: flex-end;
+  }
+
+  .mosaic-filter-row {
+    grid-template-columns: 1fr;
+  }
+
+  .mosaic-card-grid {
+    grid-template-columns: repeat(auto-fill, minmax(100px, 1fr));
+  }
+}
